refactor: EventController 통합
