+++ event-service/src/main/java/com/ojosama/eventservice/event/presentation/controller/EventController.java
@@ -4,33 +4,46 @@ import com.ojosama.common.exception.CommonErrorCode;
 import com.ojosama.common.exception.CustomException;
 import com.ojosama.common.response.ApiResponse;
 import com.ojosama.eventservice.event.application.dto.command.CreateEventCommand;
+import com.ojosama.eventservice.event.application.dto.command.EventListCommand;
 import com.ojosama.eventservice.event.application.dto.command.UpdateEventCommand;
 import com.ojosama.eventservice.event.application.dto.result.EventResult;
 import com.ojosama.eventservice.event.application.service.EventCommandService;
+import com.ojosama.eventservice.event.application.service.EventQueryService;
+import com.ojosama.eventservice.event.domain.model.EventStatus;
 import com.ojosama.eventservice.event.presentation.dto.request.CreateEventRequest;
 import com.ojosama.eventservice.event.presentation.dto.request.UpdateEventRequest;
+import com.ojosama.eventservice.event.presentation.dto.response.EventDetailResponse;
 import com.ojosama.eventservice.event.presentation.dto.response.EventResponse;
 import jakarta.validation.Valid;
+import java.time.LocalDateTime;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.web.PageableDefault;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/v1/events")
-public class EventCommandController {
+public class EventController {
 
     private final EventCommandService eventCommandService;
+    private final EventQueryService eventQueryService;
 
     @PostMapping
 //    @PreAuthorize("hasAnyRole('ADMIN', 'CONCERT_MANAGER', 'FESTIVAL_MANAGER', 'FANMEETING_MANAGER', 'POPUP_MANAGER')")
@@ -43,6 +56,35 @@ public class EventCommandController {
         EventResult result = eventCommandService.createEvent(CreateEventCommand.from(userId, request));
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(ApiResponse.created(EventResponse.from(result)));
+    }
+
+    @GetMapping
+    public ResponseEntity<ApiResponse<Page<EventResponse>>> getEvents(
+            @RequestHeader(value = "X-User-Id", required = false) UUID userId,
+            @RequestHeader(value = "X-User-Role", required = false) String userRole,
+            @RequestParam(required = false) String category,
+            @RequestParam(required = false) EventStatus status,
+            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime startAt,
+            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime endAt,
+            @RequestParam(required = false) Integer year,
+            @RequestParam(required = false) Integer month,
+            @PageableDefault(size = 10, sort = "eventTime.startAt", direction = Sort.Direction.ASC) Pageable pageable) {
+
+        validateAuthHeaders(userId, userRole);
+        EventListCommand command = new EventListCommand(category, status, startAt, endAt, year, month);
+        Page<EventResult> result = eventQueryService.getEvents(command, pageable);
+        return ResponseEntity.ok(ApiResponse.success(result.map(EventResponse::from)));
+    }
+
+    @GetMapping("/{eventId}")
+    public ResponseEntity<ApiResponse<EventDetailResponse>> getEvent(
+            @RequestHeader(value = "X-User-Id", required = false) UUID userId,
+            @RequestHeader(value = "X-User-Role", required = false) String userRole,
+            @PathVariable UUID eventId) {
+
+        validateAuthHeaders(userId, userRole);
+        EventResult result = eventQueryService.getEventById(eventId);
+        return ResponseEntity.ok(ApiResponse.success(EventDetailResponse.from(result)));
     }
 
     @PatchMapping("/{eventId}")
@@ -68,5 +110,9 @@ public class EventCommandController {
 //        if (userId == null || userRole == null) { throw new CustomException(CommonErrorCode.INVALID_TOKEN); }
         eventCommandService.deleteEvent(userId, eventId);
         return ResponseEntity.noContent().build();
+    }
+
+    private void validateAuthHeaders(UUID userId, String userRole) {
+//        if (userId == null || !StringUtils.hasText(userRole)) { throw new CustomException(CommonErrorCode.INVALID_TOKEN); }
     }
 }
