+++ event-service/src/test/java/com/ojosama/eventservice/event/presentation/controller/EventControllerTest.java
@@ -4,6 +4,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.willThrow;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
@@ -15,6 +16,7 @@ import com.ojosama.common.exception.GlobalExceptionHandler;
 import com.ojosama.eventservice.event.application.dto.result.EventResult;
 import com.ojosama.eventservice.event.application.dto.result.ScheduleResult;
 import com.ojosama.eventservice.event.application.service.EventCommandService;
+import com.ojosama.eventservice.event.application.service.EventQueryService;
 import com.ojosama.eventservice.event.domain.exception.EventErrorCode;
 import com.ojosama.eventservice.event.domain.exception.EventException;
 import com.ojosama.eventservice.event.presentation.dto.request.CreateEventRequest;
@@ -32,16 +34,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.context.annotation.Import;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
-@WebMvcTest(EventCommandController.class)
+@WebMvcTest(EventController.class)
 @AutoConfigureMockMvc(addFilters = false)
 @Import(GlobalExceptionHandler.class)
-@DisplayName("EventCommandController 테스트")
-class EventCommandControllerTest {
+@DisplayName("EventController 테스트")
+class EventControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
@@ -51,6 +55,9 @@ class EventCommandControllerTest {
 
     @MockitoBean
     private EventCommandService eventCommandService;
+
+    @MockitoBean
+    private EventQueryService eventQueryService;
 
     private ObjectMapper objectMapper;
 
@@ -67,11 +74,31 @@ class EventCommandControllerTest {
         objectMapper.registerModule(new JavaTimeModule());
     }
 
+    private EventResult buildEventResult(boolean hasTicketing) {
+        List<ScheduleResult> schedules = List.of(
+                new ScheduleResult(UUID.randomUUID(), "메인 공연", FUTURE_START, FUTURE_START.plusHours(2))
+        );
+        return new EventResult(
+                UUID.randomUUID(), "서울 재즈 페스티벌",
+                CATEGORY_ID, "FESTIVAL",
+                FUTURE_START, FUTURE_END,
+                "올림픽공원", new BigDecimal("37.52"), new BigDecimal("127.12"),
+                0, 50000,
+                hasTicketing,
+                hasTicketing ? FUTURE_START.minusDays(10) : null,
+                hasTicketing ? FUTURE_START.minusDays(1) : null,
+                hasTicketing ? "http://ticket.example.com" : null,
+                "http://official.example.com",
+                "최고의 재즈 페스티벌", "재즈밴드A", "http://img.example.com/banner.jpg",
+                "SCHEDULED",
+                schedules
+        );
+    }
+
     private CreateEventRequest buildValidCreateRequest(boolean hasTicketing) {
         List<CreateScheduleRequest> schedules = List.of(
                 new CreateScheduleRequest("메인 공연", FUTURE_START, FUTURE_START.plusHours(2))
         );
-
         if (hasTicketing) {
             return new CreateEventRequest(
                     "서울 재즈 페스티벌", CATEGORY_ID,
@@ -86,7 +113,6 @@ class EventCommandControllerTest {
                     schedules
             );
         }
-
         return new CreateEventRequest(
                 "서울 재즈 페스티벌", CATEGORY_ID,
                 FUTURE_START, FUTURE_END,
@@ -103,7 +129,6 @@ class EventCommandControllerTest {
         List<CreateScheduleRequest> schedules = withSchedules
                 ? List.of(new CreateScheduleRequest("수정된 공연", FUTURE_START, FUTURE_START.plusHours(3)))
                 : null;
-
         if (hasTicketing) {
             return new UpdateEventRequest(
                     "수정된 재즈 페스티벌", CATEGORY_ID,
@@ -118,7 +143,6 @@ class EventCommandControllerTest {
                     schedules
             );
         }
-
         return new UpdateEventRequest(
                 "수정된 재즈 페스티벌", CATEGORY_ID,
                 FUTURE_START, FUTURE_END,
@@ -127,28 +151,6 @@ class EventCommandControllerTest {
                 false, null, null, null,
                 "http://official.example.com",
                 "수정된 설명", "재즈밴드B", "http://img.example.com/updated.jpg",
-                schedules
-        );
-    }
-
-    private EventResult buildEventResult(boolean hasTicketing) {
-        List<ScheduleResult> schedules = List.of(
-                new ScheduleResult(UUID.randomUUID(), "메인 공연", FUTURE_START, FUTURE_START.plusHours(2))
-        );
-
-        return new EventResult(
-                UUID.randomUUID(), "서울 재즈 페스티벌",
-                CATEGORY_ID, "FESTIVAL",
-                FUTURE_START, FUTURE_END,
-                "올림픽공원", new BigDecimal("37.52"), new BigDecimal("127.12"),
-                0, 50000,
-                hasTicketing,
-                hasTicketing ? FUTURE_START.minusDays(10) : null,
-                hasTicketing ? FUTURE_START.minusDays(1) : null,
-                hasTicketing ? "http://ticket.example.com" : null,
-                "http://official.example.com",
-                "최고의 재즈 페스티벌", "재즈밴드A", "http://img.example.com/banner.jpg",
-                "SCHEDULED",
                 schedules
         );
     }
@@ -189,7 +191,6 @@ class EventCommandControllerTest {
                     "설명", null, "http://img.example.com/banner.jpg",
                     List.of(new CreateScheduleRequest("공연", FUTURE_START, FUTURE_START.plusHours(2)))
             );
-
             mockMvc.perform(post("/v1/events")
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE)
@@ -210,7 +211,6 @@ class EventCommandControllerTest {
                     "설명", null, "http://img.example.com/banner.jpg",
                     List.of()
             );
-
             mockMvc.perform(post("/v1/events")
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE)
@@ -224,7 +224,6 @@ class EventCommandControllerTest {
         void createEvent_ticketingTrueWithoutLink_returns400() throws Exception {
             given(eventCommandService.createEvent(any()))
                     .willThrow(new EventException(EventErrorCode.TICKETING_NOT_AVAILABLE));
-
             mockMvc.perform(post("/v1/events")
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE)
@@ -238,7 +237,6 @@ class EventCommandControllerTest {
         void createEvent_startAfterEnd_returns400() throws Exception {
             given(eventCommandService.createEvent(any()))
                     .willThrow(new EventException(EventErrorCode.EVENT_INVALID_TIME));
-
             mockMvc.perform(post("/v1/events")
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE)
@@ -252,7 +250,6 @@ class EventCommandControllerTest {
         void createEvent_unknownCategory_returns404() throws Exception {
             given(eventCommandService.createEvent(any()))
                     .willThrow(new EventException(EventErrorCode.EVENT_CATEGORY_NOT_FOUND));
-
             mockMvc.perform(post("/v1/events")
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE)
@@ -270,7 +267,6 @@ class EventCommandControllerTest {
         @DisplayName("티켓팅 없는 행사 등록 → 201, SCHEDULED")
         void createEvent_withoutTicketing_returns201() throws Exception {
             given(eventCommandService.createEvent(any())).willReturn(buildEventResult(false));
-
             mockMvc.perform(post("/v1/events")
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE)
@@ -288,7 +284,6 @@ class EventCommandControllerTest {
         @DisplayName("티켓팅 있는 행사 등록 → 201")
         void createEvent_withTicketing_returns201() throws Exception {
             given(eventCommandService.createEvent(any())).willReturn(buildEventResult(true));
-
             mockMvc.perform(post("/v1/events")
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE)
@@ -299,6 +294,113 @@ class EventCommandControllerTest {
                     .andExpect(jsonPath("$.data.status").value("SCHEDULED"))
                     .andExpect(jsonPath("$.data.hasTicketing").value(true))
                     .andExpect(jsonPath("$.data.ticketingLink").value("http://ticket.example.com"));
+        }
+    }
+
+    @Nested
+    @DisplayName("행사 목록 조회 실패")
+    class GetEventsFailure {
+
+        @Test
+        @DisplayName("X-User-Id 미전달 → 401")
+        void getEvents_missingUserId_returns401() throws Exception {
+            mockMvc.perform(get("/v1/events")
+                            .header("X-User-Role", MANAGER_ROLE))
+                    .andExpect(status().isUnauthorized());
+        }
+
+        @Test
+        @DisplayName("X-User-Role 미전달 → 401")
+        void getEvents_missingUserRole_returns401() throws Exception {
+            mockMvc.perform(get("/v1/events")
+                            .header("X-User-Id", USER_ID.toString()))
+                    .andExpect(status().isUnauthorized());
+        }
+    }
+
+    @Nested
+    @DisplayName("행사 목록 조회 성공")
+    class GetEventsSuccess {
+
+        @Test
+        @DisplayName("전체 조회 → 200, 페이지 정보 포함")
+        void getEvents_noFilter_returns200WithPage() throws Exception {
+            given(eventQueryService.getEvents(any(), any()))
+                    .willReturn(new PageImpl<>(List.of(buildEventResult(false)), PageRequest.of(0, 10), 1));
+            mockMvc.perform(get("/v1/events")
+                            .header("X-User-Id", USER_ID.toString())
+                            .header("X-User-Role", MANAGER_ROLE))
+                    .andExpect(status().isOk())
+                    .andExpect(jsonPath("$.status").value(200))
+                    .andExpect(jsonPath("$.data.content").isArray())
+                    .andExpect(jsonPath("$.data.content.length()").value(1))
+                    .andExpect(jsonPath("$.data.totalElements").value(1))
+                    .andExpect(jsonPath("$.data.totalPages").value(1));
+        }
+
+        @Test
+        @DisplayName("조회 결과 없을 때 빈 목록 반환 → 200")
+        void getEvents_noResult_returnsEmptyContent() throws Exception {
+            given(eventQueryService.getEvents(any(), any()))
+                    .willReturn(new PageImpl<>(List.of(), PageRequest.of(0, 10), 0));
+            mockMvc.perform(get("/v1/events")
+                            .header("X-User-Id", USER_ID.toString())
+                            .header("X-User-Role", MANAGER_ROLE))
+                    .andExpect(status().isOk())
+                    .andExpect(jsonPath("$.data.content").isArray())
+                    .andExpect(jsonPath("$.data.content.length()").value(0));
+        }
+    }
+
+    @Nested
+    @DisplayName("행사 상세 조회 실패")
+    class GetEventFailure {
+
+        @Test
+        @DisplayName("X-User-Id 미전달 → 401")
+        void getEvent_missingUserId_returns401() throws Exception {
+            mockMvc.perform(get("/v1/events/{eventId}", UUID.randomUUID())
+                            .header("X-User-Role", MANAGER_ROLE))
+                    .andExpect(status().isUnauthorized());
+        }
+
+        @Test
+        @DisplayName("X-User-Role 미전달 → 401")
+        void getEvent_missingUserRole_returns401() throws Exception {
+            mockMvc.perform(get("/v1/events/{eventId}", UUID.randomUUID())
+                            .header("X-User-Id", USER_ID.toString()))
+                    .andExpect(status().isUnauthorized());
+        }
+
+        @Test
+        @DisplayName("존재하지 않는 행사 ID → 404")
+        void getEvent_unknownEventId_returns404() throws Exception {
+            given(eventQueryService.getEventById(any()))
+                    .willThrow(new EventException(EventErrorCode.EVENT_NOT_FOUND));
+            mockMvc.perform(get("/v1/events/{eventId}", UUID.randomUUID())
+                            .header("X-User-Id", USER_ID.toString())
+                            .header("X-User-Role", MANAGER_ROLE))
+                    .andExpect(status().isNotFound());
+        }
+    }
+
+    @Nested
+    @DisplayName("행사 상세 조회 성공")
+    class GetEventSuccess {
+
+        @Test
+        @DisplayName("정상 조회 → 200, schedules 포함")
+        void getEvent_validEventId_returns200() throws Exception {
+            UUID eventId = UUID.randomUUID();
+            given(eventQueryService.getEventById(eventId)).willReturn(buildEventResult(false));
+            mockMvc.perform(get("/v1/events/{eventId}", eventId)
+                            .header("X-User-Id", USER_ID.toString())
+                            .header("X-User-Role", MANAGER_ROLE))
+                    .andExpect(status().isOk())
+                    .andExpect(jsonPath("$.status").value(200))
+                    .andExpect(jsonPath("$.data.status").value("SCHEDULED"))
+                    .andExpect(jsonPath("$.data.schedules").isArray())
+                    .andExpect(jsonPath("$.data.schedules.length()").value(1));
         }
     }
 
@@ -338,7 +440,6 @@ class EventCommandControllerTest {
                     "설명", null, "http://img.example.com/banner.jpg",
                     List.of(new CreateScheduleRequest("공연", FUTURE_START, FUTURE_START.plusHours(2)))
             );
-
             mockMvc.perform(patch("/v1/events/{eventId}", UUID.randomUUID())
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE)
@@ -352,7 +453,6 @@ class EventCommandControllerTest {
         void updateEvent_ticketingTrueWithoutLink_returns400() throws Exception {
             given(eventCommandService.updateEvent(any()))
                     .willThrow(new EventException(EventErrorCode.TICKETING_NOT_AVAILABLE));
-
             mockMvc.perform(patch("/v1/events/{eventId}", UUID.randomUUID())
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE)
@@ -366,7 +466,6 @@ class EventCommandControllerTest {
         void updateEvent_unknownEventId_returns404() throws Exception {
             given(eventCommandService.updateEvent(any()))
                     .willThrow(new EventException(EventErrorCode.EVENT_NOT_FOUND));
-
             mockMvc.perform(patch("/v1/events/{eventId}", UUID.randomUUID())
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE)
@@ -384,7 +483,6 @@ class EventCommandControllerTest {
         @DisplayName("schedules 포함 → 200, 일정 교체됨")
         void updateEvent_withSchedules_returns200() throws Exception {
             given(eventCommandService.updateEvent(any())).willReturn(buildEventResult(false));
-
             mockMvc.perform(patch("/v1/events/{eventId}", UUID.randomUUID())
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE)
@@ -401,14 +499,12 @@ class EventCommandControllerTest {
         @DisplayName("schedules 생략 → 200, 기존 일정 유지")
         void updateEvent_withoutSchedules_returns200() throws Exception {
             given(eventCommandService.updateEvent(any())).willReturn(buildEventResult(false));
-
             mockMvc.perform(patch("/v1/events/{eventId}", UUID.randomUUID())
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE)
                             .contentType(MediaType.APPLICATION_JSON)
                             .content(objectMapper.writeValueAsString(buildValidUpdateRequest(false, false))))
                     .andExpect(status().isOk())
-                    .andExpect(jsonPath("$.status").value(200))
                     .andExpect(jsonPath("$.data.schedules").isArray());
         }
 
@@ -416,7 +512,6 @@ class EventCommandControllerTest {
         @DisplayName("티켓팅 있는 행사 수정 → 200")
         void updateEvent_withTicketing_returns200() throws Exception {
             given(eventCommandService.updateEvent(any())).willReturn(buildEventResult(true));
-
             mockMvc.perform(patch("/v1/events/{eventId}", UUID.randomUUID())
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE)
@@ -453,7 +548,6 @@ class EventCommandControllerTest {
         void deleteEvent_unknownEventId_returns404() throws Exception {
             willThrow(new EventException(EventErrorCode.EVENT_NOT_FOUND))
                     .given(eventCommandService).deleteEvent(any(), any());
-
             mockMvc.perform(delete("/v1/events/{eventId}", UUID.randomUUID())
                             .header("X-User-Id", USER_ID.toString())
                             .header("X-User-Role", MANAGER_ROLE))
